requirements.txt
tensorflow-cpu
flask
underthesea